+++ engine/launcher/src/main/scala/org/enso/launcher/OS.scala
@@ -1,6 +1,4 @@
-package org.enso.launcher.internal
-
-import org.enso.launcher.Logger
+package org.enso.launcher
 
 sealed trait OS
 object OS {
