--- engine/launcher/src/main/scala/org/enso/launcher/PluginManager.scala
@@ -1,7 +1,0 @@
-package org.enso.launcher
-
-/**
-  * Default implementation of the [[internal.PluginManager]] using the default
-  * [[Environment]].
-  */
-object PluginManager extends internal.PluginManager(Environment)
